Generate base request

This function takes a url and creates a basic httr2 request that adds the user-agent and adds an authorization token to the X-Esri-Authorization header.

Usage

arc_base_req(
  url,
  token = NULL,
  path = NULL,
  query = NULL,
  error_call = rlang::caller_env()
)

Arguments

url

a valid url that is passed to httr2::request()

token

an object of class httr2_token as generated by auth_code() or related function

path

a character vector of paths to be appended to url using httr2::req_url_path_append()

query

a named vector or named list of query parameters to be appended to the url using httr2::req_url_query()

error_call

the caller environment to be used when propagating errors.

Value

an httr2_request with the X-Esri-Authorization header and User-Agent set.

Examples

arc_base_req("https://arcgis.com")
#> <httr2_request>
#> GET https://arcgis.com
#> Body: empty
#> Options:
#> * useragent: "arcgisutils v0.4.0"