Fetch Group Information

Fetches metadata about a group based on a provided group_id.

Usage

arc_group(group_id, host = arc_host(), token = arc_token())

Arguments

group_id

the unique group identifier. A scalar character.

host

default "https://www.arcgis.com"

token

an httr2_token as created by auth_code() or similar

Value

a list with group metadata

Details

See also

Examples

arc_group("2f0ec8cb03574128bd673cefab106f39")
#> $id
#> [1] "2f0ec8cb03574128bd673cefab106f39"
#> 
#> $title
#> [1] "Web Application Templates"
#> 
#> $isInvitationOnly
#> [1] TRUE
#> 
#> $owner
#> [1] "esri_en"
#> 
#> $description
#> [1] "<span style='color:rgb(36, 41, 47); font-family:-apple-system, BlinkMacSystemFont, &quot;Segoe UI&quot;, Helvetica, Arial, sans-serif, &quot;Apple Color Emoji&quot;, &quot;Segoe UI Emoji&quot;; font-size:14px; font-style:normal; font-variant-ligatures:normal; font-variant-caps:normal; font-weight:400; letter-spacing:normal; text-align:start; text-indent:0px; text-transform:none; word-spacing:0px; background-color:rgb(255, 255, 255); text-decoration-style:initial; text-decoration-color:initial; display:inline !important;'>Make it easy for your audience to engage with your maps and data by sharing them using configurable app templates such as ArcGIS Instant Apps. In a few minutes, you can select a template, choose configuration options, and share a web app that addresses your audience's needs. Instant Apps are built to support accessibility across audiences and devices, and are included with ArcGIS Online, ArcGIS Pro, and ArcGIS Enterprise.</span>"
#> 
#> $snippet
#> [1] "Templates for publishing your web map application."
#> 
#> $tags
#> [1] "Esri"      "Colors"    "templates"
#> 
#> $typeKeywords
#> NULL
#> 
#> $phone
#> [1] ""
#> 
#> $sortField
#> NULL
#> 
#> $sortOrder
#> NULL
#> 
#> $isViewOnly
#> [1] FALSE
#> 
#> $featuredItemsId
#> NULL
#> 
#> $thumbnail
#> [1] "thumbnail1679686061624.png"
#> 
#> $created
#> [1] 1.328834e+12
#> 
#> $modified
#> [1] 1.679686e+12
#> 
#> $access
#> [1] "public"
#> 
#> $capabilities
#> NULL
#> 
#> $isFav
#> [1] FALSE
#> 
#> $isReadOnly
#> [1] FALSE
#> 
#> $protected
#> [1] FALSE
#> 
#> $autoJoin
#> [1] FALSE
#> 
#> $notificationsEnabled
#> [1] FALSE
#> 
#> $provider
#> NULL
#> 
#> $providerGroupName
#> NULL
#> 
#> $leavingDisallowed
#> [1] FALSE
#> 
#> $hiddenMembers
#> [1] FALSE
#> 
#> $displaySettings
#> $displaySettings$itemTypes
#> [1] ""
#> 
#> 
#> $orgId
#> [1] "none"
#> 
#> $properties
#> NULL
#> 
#> $robots
#> [1] "noindex"  "nofollow"
#>