List users in a group

List all users in a given group.

Usage

arc_group_users(
  group,
  name = NULL,
  member_type = NULL,
  joined_before = NULL,
  joined_after = NULL,
  sort_field = c("username", "membertype", "joined"),
  sort_order = c("asc", "desc"),
  page_size = 50,
  max_pages = Inf,
  .progress = TRUE,
  host = arc_host(),
  token = arc_token()
)

Arguments

group

a group ID as a scalar string or PortalGroup created via arc_group()

name

a scalar string of a user to search for.

member_type

default NULL. If provided must be one of "admin" or "member".

joined_before

default NULL. A scalar date to search for users who joined before this date.

joined_after

default NULL. A scalar date to search for users who joined after this date.

sort_field

default "username". The field to sort by. Must be one of "username", "membertype", or "joined".

sort_order

optional string. One of either asc or desc for ascending or descending order respectively.

page_size

a scalar integer between 1 and 100 indicating the number of responses per page.

max_pages

the maximum number of pages to fetch. By default fetches all pages.

.progress

default TRUE. Whether to display a progress bar for requests.

host

default "https://www.arcgis.com". The host of your ArcGIS Portal.

token

an object of class httr2_token as generated by auth_code() or related function

Value

a data.frame

References

API Reference

Examples

if (FALSE) { # \dontrun{
set_arc_token(auth_user())
groups <- arc_user("r-bridge-docs")$groups
arc_group_users(groups$id[1])
} # }