Access the Self Resource

The function returns the /self resource from the ArcGIS REST API. The /self endpoint returns the view of the portal as seen by the current user, whether anonymous or signed in.

Usage

arc_self_meta(token = arc_token(), error_call = rlang::current_call())

Arguments

token

an object of class httr2_token as generated by auth_code() or related function

error_call

the caller environment to be used when propagating errors.

Value

A named list.

Details

See the endpoint documentation for more details.

The Portal Self response can vary based on whether it's called by a user, an app, or both.

The response includes user and appinfo properties, and the variations in responses are primarily related to these two properties. As the names indicate, the user property includes information about the user making the call, and the appinfo property includes information pertaining to the app that made the call.

Examples

if (FALSE) {
set_arc_token(auth_code())
self <- arc_self_meta()
names(self)
}