Retrieve metadata

Utility functions for feature service metadata.

Usage

fetch_layer_metadata(url, token = NULL, call = rlang::caller_env())

Arguments

url

the url of the item.

token

an httr2_token from one of the provided auth_ functions

call

default rlang::caller_env(). The calling environment passed to detect_errors().

Value

returns a list object

Details

fetch_layer_metadata() given a request, fetches the metadata by setting the query parameter f=json

Examples

# url is broken into parts to fit within 100 characters to avoid CRAN notes
url_parts <- c(
  "https://services.arcgis.com/P3ePLMYs2RVChkJx/ArcGIS/rest/services",
  "/USA_Counties_Generalized_Boundaries/FeatureServer/0"
)

furl <- paste0(url_parts, collapse = "")
meta <- fetch_layer_metadata(furl)
head(names(meta))
#> [1] "currentVersion" "id"             "name"           "type"          
#> [5] "serviceItemId"  "cacheMaxAge"